 DFR Test Automation Suite/dfr_testing/automation_suite/integration_tests_odoo/common/odoo_test_helper.py
```python
import odoorpc # Example library for external RPC connection
import os
from dfr.testing.automation.utils.logger_setup import setup_logger
from dfr.testing.automation.config import global_config # For target env
# from dfr.testing.automation.utils.config_loader import ConfigLoader # For DB details via INI files

# Note: Using Odoo's native test framework (e.g., inheriting from odoo.tests.common.TransactionCase)
# is strongly recommended for most Odoo integration tests. It provides proper test environment
# setup, database transaction management (rollback after each test), and direct access to `self.env`.
# This OdooTestHelper class is more illustrative of how one *might* interact with Odoo externally
# via RPC for certain types of integration tests, or if tests must run outside the Odoo process.
# Full transaction management for external tests is complex and not handled by this basic helper.

logger = setup_logger(__name__)

class OdooTestHelper:
    """
    Utility class for interacting with an Odoo instance, primarily for integration tests
    that might run externally to the Odoo process.
    
    It can be initialized with an existing Odoo environment object (e.g., from Odoo's
    native test runner if this helper is adapted) or attempt an RPC connection.
    """
    def __init__(self, odoo_env=None, rpc_config: dict | None = None):
        """
        Initializes the OdooTestHelper.

        Args:
            odoo_env: An Odoo `env` object if running within the Odoo test framework or if
                      an RPC client providing such an object is already established.
            rpc_config (dict, optional): Configuration for RPC connection if `odoo_env` is not provided.
                                         Expected keys: 'host', 'port', 'db_name', 'user', 'password'.
                                         Password can be actual password or env var name if logic is added to resolve it.
        """
        self.env = odoo_env
        self.odoo_rpc_client = None # Holds the odoorpc.ODOO instance if connected

        if not self.env and rpc_config:
            logger.info("Odoo environment (self.env) not provided. Attempting to initialize Odoo RPC client.")
            try:
                self._connect_odoo_rpc(rpc_config)
                if self.odoo_rpc_client and hasattr(self.odoo_rpc_client, 'env'):
                    self.env = self.odoo_rpc_client.env # odoorpc.ODOO object behaves like a remote env
                else:
                    logger.error("Failed to establish Odoo environment via RPC client.")
            except Exception as e:
                logger.error(f"Failed to connect to Odoo RPC: {e}")
                self.env = None
        
        if not self.env:
            logger.warning("Odoo environment is not available. Helper functionality will be limited or may fail.")
        else:
            logger.info("OdooTestHelper initialized with an Odoo environment.")


    def _connect_odoo_rpc(self, rpc_config: dict):
        """
        Illustrative method to connect to a running Odoo instance via RPC using `odoorpc`.
        This requires Odoo's XML-RPC service to be enabled and accessible.

        Args:
            rpc_config (dict): Connection details ('host', 'port', 'db_name', 'user', 'password').
        
        Raises:
            ConnectionError: If connection or login fails.
            KeyError: If rpc_config is missing required keys.
        """
        try:
            host = rpc_config['host']
            port = int(rpc_config['port'])
            db_name = rpc_config['db_name']
            user = rpc_config['user']
            # Password might be direct or an env var name. For simplicity, assume direct here.
            # In a real scenario, use ConfigLoader.get_env_credential if it's a ref.
            password = rpc_config['password'] 

            if not all([host, port, db_name, user, password]):
                logger.error("Odoo RPC connection details are incomplete in provided rpc_config.")
                raise ValueError("Incomplete Odoo RPC connection details.")

            logger.info(f"Attempting to connect to Odoo RPC at {host}:{port} for DB '{db_name}' as user '{user}'...")
            
            self.odoo_rpc_client = odoorpc.ODOO(host, port=port)
            self.odoo_rpc_client.login(db_name, user, password)
            
            logger.info("Odoo RPC connection and login successful.")
            
        except KeyError as e:
            logger.error(f"Missing key in rpc_config for Odoo RPC connection: {e}")
            self.odoo_rpc_client = None
            raise
        except (odoorpc.error.RPCError, odoorpc.error.ConnectorError) as e:
            logger.error(f"Odoo RPC connection or login failed: {e}")
            self.odoo_rpc_client = None
            raise ConnectionError(f"Odoo RPC failure: {e}")
        except Exception as e:
            logger.error(f"An unexpected error occurred during Odoo RPC connection: {e}")
            self.odoo_rpc_client = None
            raise ConnectionError(f"Unexpected Odoo RPC error: {e}")


    def create_test_farmer(self, name: str, uid: str | None = None, **kwargs):
        """
        Creates a test farmer record directly via Odoo ORM (using `self.env`).

        Args:
            name (str): The full name of the farmer.
            uid (str, optional): The UID for the farmer. If None, Odoo might generate one or it might be required.
            **kwargs: Additional field values for the farmer record (e.g., national_id_number).

        Returns:
            An Odoo record object representing the created farmer, or None/raises Exception on failure.
            The exact type depends on whether self.env is from native Odoo or odoorpc.

        Raises:
            RuntimeError: If Odoo environment (`self.env`) is not available.
            Exception: If Odoo record creation fails (e.g., validation error, database error).
        """
        if not self.env:
            msg = "Odoo environment not available for create_test_farmer."
            logger.error(msg)
            raise RuntimeError(msg)

        # Ensure 'dfr.farmer' is the correct Odoo model name for farmers
        farmer_model = self.env['dfr.farmer'] 
        
        vals = {'name': name, **kwargs}
        if uid:
            vals['uid'] = uid
        # Ensure all required fields by the Odoo model are provided in vals or have defaults in Odoo

        logger.info(f"Attempting to create test farmer with values: {vals}")
        try:
            farmer_record = farmer_model.create(vals)
            # For external RPC, changes are often auto-committed or might need explicit commit
            # depending on odoorpc version and server config.
            # If using this helper within an Odoo native test (TransactionCase), commits are handled by the framework.
            # Example: if self.env.cr and hasattr(self.env.cr, 'commit'): self.env.cr.commit()

            # Log the ID and UID (if available on the record object)
            created_id = getattr(farmer_record, 'id', 'N/A')
            created_uid = getattr(farmer_record, 'uid', 'N/A') # Assuming 'uid' field exists
            logger.info(f"Successfully created test farmer: {name} (ID: {created_id}, UID: {created_uid})")
            return farmer_record
        except Exception as e: # Catch Odoo ORM exceptions (e.g., ValidationError, IntegrityError)
            logger.error(f"Failed to create test farmer '{name}': {e}")
            # If using external RPC with manual transaction needs:
            # Example: if self.env.cr and hasattr(self.env.cr, 'rollback'): self.env.cr.rollback()
            raise # Re-raise the exception for the test framework to handle

    def get_farmer_by_uid(self, uid: str):
        """
        Retrieves a farmer record by UID via Odoo ORM.

        Args:
            uid (str): The UID of the farmer to search for.

        Returns:
            An Odoo record object (or recordset) if found, otherwise an empty recordset or None.
        
        Raises:
            RuntimeError: If Odoo environment (`self.env`) is not available.
        """
        if not self.env:
            msg = "Odoo environment not available for get_farmer_by_uid."
            logger.error(msg)
            raise RuntimeError(msg)

        farmer_model = self.env['dfr.farmer']
        logger.info(f"Searching for farmer with UID: {uid}")
        # search() returns a recordset; limit=1 gets at most one record.
        # The result is a recordset, even if empty or with one record.
        farmers = farmer_model.search([('uid', '=', uid)], limit=1) 
        
        if farmers:
            logger.info(f"Found farmer with UID {uid} (ID: {getattr(farmers, 'id', 'N/A')}).")
            return farmers # Returns a recordset (possibly single item)
        else:
            logger.info(f"Farmer with UID {uid} not found.")
            return farmers # Returns an empty recordset

    def get_service(self, service_name: str):
        """
        Hypothetical method to access a custom Odoo service.
        The implementation is highly dependent on how services are defined and exposed
        in the DFR Odoo modules (e.g., via `self.env['ir.actions.server']` or custom registry).
        
        Args:
            service_name (str): The name of the service to access.

        Returns:
            The service object or result of a service call, or None if not found/applicable.
        """
        if not self.env:
            msg = "Odoo environment not available for get_service."
            logger.error(msg)
            raise RuntimeError(msg)
            
        logger.warning(f"get_service for '{service_name}' is a placeholder. "
                       "Actual implementation depends on DFR Odoo module design.")
        # Example if services were registered in a specific way:
        # if service_name == 'dfr.deduplication.service':
        #     return self.env['dfr.deduplication.service'] # Hypothetical model name for service
        return None

    # Add more helper methods as needed for specific Odoo integration scenarios:
    # - create_household(data)
    # - create_plot(farm_id, data)
    # - trigger_odoo_workflow(model_name, record_id, workflow_signal)
    # - check_record_exists(model_name, domain)
    # - delete_record(record_object)
```