 DFR Test Automation Suite/dfr_testing/automation_suite/requirements.txt
pytest==8.2.2
pytest-html==4.1.1
pytest-xdist==3.6.1
pytest-benchmark==4.0.0
selenium==4.21.0
Appium-Python-Client==4.0.0
requests==2.32.3
jsonschema==4.22.0
configparser==7.0.0
PyYAML==6.0.1
webdriver-manager==4.0.1 # For easy management of browser drivers